requirements.txt
cmake==3.17.3
dlib==19.20.0
cvlib==0.2.5
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==1.1.1
gunicorn==20.0.4
Jinja2==2.11.1
opencv-contrib-python==4.3.0.36
opencv-python==4.2.0.34
Pillow==7.0.0
unicodecsv==0.14.1
virtualenv==20.0.25
Werkzeug==1.0.0
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
win-unicode-console==0.5

